cleanup model location
